backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
mysqlclient==2.2.0
psycopg2-binary==2.9.9
google-analytics-data==0.18.0
google-api-python-client==2.108.0
google-auth==2.26.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dateutil==2.8.2
requests==2.32.3
